Implement delete cards functionality

+++ src/Flashcards.WebAPI/Flashcards.Core/DTO/AffectedResponse.cs
@@ -2,6 +2,6 @@
 {
 	public class AffectedResponse
 	{
-		public int Affected { get; set; }
+		public long Affected { get; set; }
 	}
 }
